requirements.txt
numpy
pandas
requests
python-dotenv
pytest
black
flake8 